requirements.txt
scipy==1.1.0
torchvision==0.2.1
numpy==1.15.2
opencv_python==3.4.3.18
torch==0.4.1
Pillow==6.2.0
